More WIP. Segfaults on sort, I think somewhere off of the main thread.

+++ config.h
@@ -24,7 +24,7 @@ int end_current_sortfile(int fd);
 void unload_midas_module();
 int pre_sort(int frag_idx, int end_idx);
 int user_addto_window(int win_strt, int new_frag);
-int default_sort(int win_idx, int frag_idx, int flag, FILE *out);
+uint8_t default_sort(int win_idx, int frag_idx, int flag, FILE *out);
 int sort_built_event(int window_start, int win_end, FILE *out);
 /////////////////////////////////////////////////////////////////////////
 /////////////////////       Histograms       ////////////////////////////
